+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
+## [Unreleased]
+
+### Fixed
+
+- Instruct agents to write Frontman feedback in English regardless of the session language.
+
 ## [5.0.0] - 2026-09-04